Move AssemblyTests
