+++ Tests/Batch1/Reflection/AssemblyTests.cs
@@ -12,6 +12,74 @@ namespace Bridge.ClientTest.Batch1.Reflection
     [TestFixture(TestNameFormat = "Reflection - Assembly {0}")]
     public class AssemblyTests
     {
+        [NonScriptable]
+        [External]
+        public class A1Attribute : Attribute
+        {
+            public int X
+            {
+                get; private set;
+            }
+
+            public int P
+            {
+                get; set;
+            }
+
+            public A1Attribute()
+            {
+            }
+
+            public A1Attribute(int x)
+            {
+                this.X = x;
+            }
+        }
+
+        public class A2Attribute : Attribute
+        {
+            public int X
+            {
+                get; private set;
+            }
+
+            public int P
+            {
+                get; set;
+            }
+
+            public A2Attribute()
+            {
+            }
+
+            public A2Attribute(int x)
+            {
+                this.X = x;
+            }
+        }
+
+        public class A3Attribute : Attribute
+        {
+            public int X
+            {
+                get; private set;
+            }
+
+            public int P
+            {
+                get; set;
+            }
+
+            public A3Attribute()
+            {
+            }
+
+            public A3Attribute(int x)
+            {
+                this.X = x;
+            }
+        }
+
         private class C
         {
         }
@@ -35,11 +103,14 @@ namespace Bridge.ClientTest.Batch1.Reflection
 	x.Foo.baz.Test = function() {};
 	return x;")]
             get
-            { return null; }
+            {
+                return null;
+            }
         }
 
         private class G<T1, T2>
-        { }
+        {
+        }
 
         private string CompilerVersion
         {
@@ -87,28 +158,28 @@ namespace Bridge.ClientTest.Batch1.Reflection
         [Test]
         public void GetExecutingAssemblyWorks()
         {
-            Assert.AreEqual(Assembly.GetExecutingAssembly().FullName, AssemblyWithVersion);
+            Assert.AreEqual(AssemblyWithVersion, Assembly.GetExecutingAssembly().FullName);
         }
 
         [Test]
         public void GetAssemblyForTypeWorks()
         {
-            Assert.AreEqual(Assembly.GetAssembly(typeof(int)).FullName, MscorlibWithVersion);
-            Assert.AreEqual(Assembly.GetAssembly(typeof(AssemblyTests)).FullName, AssemblyWithVersion);
+            Assert.AreEqual(MscorlibWithVersion, Assembly.GetAssembly(typeof(int)).FullName);
+            Assert.AreEqual(AssemblyWithVersion, Assembly.GetAssembly(typeof(AssemblyTests)).FullName);
         }
 
         [Test]
         public void FullNameWorks()
         {
-            Assert.AreEqual(typeof(int).Assembly.FullName, MscorlibWithVersion);
-            Assert.AreEqual(typeof(AssemblyTests).Assembly.FullName, AssemblyWithVersion);
+            Assert.AreEqual(MscorlibWithVersion, typeof(int).Assembly.FullName);
+            Assert.AreEqual(AssemblyWithVersion, typeof(AssemblyTests).Assembly.FullName);
         }
 
         [Test]
         public void ToStringWorks()
         {
-            Assert.AreEqual(typeof(int).Assembly.ToString(), MscorlibWithVersion);
-            Assert.AreEqual(typeof(AssemblyTests).Assembly.ToString(), AssemblyWithVersion);
+            Assert.AreEqual(MscorlibWithVersion, typeof(int).Assembly.ToString());
+            Assert.AreEqual(AssemblyWithVersion, typeof(AssemblyTests).Assembly.ToString());
         }
 
         [Test]
@@ -126,7 +197,7 @@ namespace Bridge.ClientTest.Batch1.Reflection
         {
             var asm = ImportedModuleTestCase;
             var types = asm.GetTypes();
-            Assert.AreEqual(types.Length, 3, "Length");
+            Assert.AreEqual(3, types.Length, "Length");
             Assert.True(types.Contains((Type)((dynamic)asm).Foo.Bar.Inner.OtherFunction), "#1");
             Assert.True(types.Contains((Type)((dynamic)asm).Foo.Bar.Something), "#2");
             Assert.True(types.Contains((Type)((dynamic)asm).Foo.baz.Test), "#3");
@@ -160,19 +231,19 @@ namespace Bridge.ClientTest.Batch1.Reflection
         [Test]
         public void AssemblyOfBuiltInTypes()
         {
-            Assert.AreEqual(typeof(DateTime).Assembly.FullName, MscorlibWithVersion);
-            Assert.AreEqual(typeof(double).Assembly.FullName, MscorlibWithVersion);
-            Assert.AreEqual(typeof(bool).Assembly.FullName, MscorlibWithVersion);
-            Assert.AreEqual(typeof(string).Assembly.FullName, MscorlibWithVersion);
-            Assert.AreEqual(typeof(Delegate).Assembly.FullName, MscorlibWithVersion);
-            Assert.AreEqual(typeof(int[]).Assembly.FullName, MscorlibWithVersion);
+            Assert.AreEqual(MscorlibWithVersion, typeof(DateTime).Assembly.FullName);
+            Assert.AreEqual(MscorlibWithVersion, typeof(double).Assembly.FullName);
+            Assert.AreEqual(MscorlibWithVersion, typeof(bool).Assembly.FullName);
+            Assert.AreEqual(MscorlibWithVersion, typeof(string).Assembly.FullName);
+            Assert.AreEqual(MscorlibWithVersion, typeof(Delegate).Assembly.FullName);
+            Assert.AreEqual(MscorlibWithVersion, typeof(int[]).Assembly.FullName);
         }
 
         [Test]
         public void CreateInstanceWorks()
         {
             Assert.True(typeof(C).Assembly.CreateInstance(typeof(C).FullName) is C, "#1");
-            Assert.AreEqual(typeof(int).Assembly.CreateInstance(typeof(int).FullName), 0, "#2");
+            Assert.AreEqual(0, typeof(int).Assembly.CreateInstance(typeof(int).FullName), "#2");
             Assert.True(typeof(C).Assembly.CreateInstance("NonExistentType") == null, "#3");
         }
 
@@ -184,19 +255,19 @@ namespace Bridge.ClientTest.Batch1.Reflection
             {
                 Assert.False(a.Some(x => x.GetType().Name == "A1Attribute"));
                 var a2 = a.Filter(x => x is AssemblyAttributes.A2Attribute);
-                Assert.AreEqual(a2.Length, 1);
+                Assert.AreEqual(1, a2.Length);
                 Assert.True(((AssemblyAttributes.A2Attribute)a2[0]).X == 64);
                 Assert.True(((AssemblyAttributes.A2Attribute)a2[0]).P == 23);
 
                 var a3 = a.Filter(x => x is AssemblyAttributes.A3Attribute);
-                Assert.AreEqual(a3.Length, 1);
+                Assert.AreEqual(1, a3.Length);
                 Assert.True(((AssemblyAttributes.A3Attribute)a3[0]).X == 15);
                 Assert.True(((AssemblyAttributes.A3Attribute)a3[0]).P == 45);
             }
 
-            foreach (var a in new[] { asm.GetCustomAttributes(typeof(AssemblyAttributes.A2Attribute)), asm.GetCustomAttributes(typeof(AssemblyAttributes.A2Attribute), true), asm.GetCustomAttributes(typeof(AssemblyAttributes.A2Attribute), false) })
+            foreach (var a in new[] { asm.GetCustomAttributes(typeof(A2Attribute)), asm.GetCustomAttributes(typeof(A2Attribute), true), asm.GetCustomAttributes(typeof(A2Attribute), false) })
             {
-                Assert.AreEqual(a.Length, 1);
+                Assert.AreEqual(1, a.Length);
                 Assert.True(((AssemblyAttributes.A2Attribute)a[0]).X == 64);
                 Assert.True(((AssemblyAttributes.A2Attribute)a[0]).P == 23);
             }
@@ -217,38 +288,38 @@ namespace Bridge.ClientTest.Batch1.Reflection
         {
             var names = Assembly.GetExecutingAssembly().GetManifestResourceNames();
             names.Sort((a, b) => a.CompareTo(b));
-            Assert.AreEqual(names, new[] { "Bridge.ClientTest.Batch1.Reflection.Resource1.bin", "Bridge.ClientTest.Batch1.Reflection.Resource2.bin" });
+            Assert.AreEqual(new[] { "Bridge.ClientTest.Batch1.Reflection.Resource1.bin", "Bridge.ClientTest.Batch1.Reflection.Resource2.bin" }, names);
         }
 
         [Test]
         public void GetManifestResourceDataAsBase64WithoutTypeWorks()
         {
-            Assert.AreEqual(Assembly.GetExecutingAssembly().GetManifestResourceDataAsBase64("Bridge.ClientTest.Batch1.Reflection.Resource1.bin"), "AAECAwQFBgc=", "#1");
-            Assert.AreEqual(Assembly.GetExecutingAssembly().GetManifestResourceDataAsBase64("Bridge.ClientTest.Batch1.Reflection.Resource2.bin"), "EBESExQV", "#2");
+            Assert.AreEqual("AAECAwQFBgc=", Assembly.GetExecutingAssembly().GetManifestResourceDataAsBase64("Bridge.ClientTest.Batch1.Reflection.Resource1.bin"), "#1");
+            Assert.AreEqual("EBESExQV", Assembly.GetExecutingAssembly().GetManifestResourceDataAsBase64("Bridge.ClientTest.Batch1.Reflection.Resource2.bin"), "#2");
             Assert.Null(Assembly.GetExecutingAssembly().GetManifestResourceDataAsBase64("NonExistent"), "#3");
         }
 
         [Test]
         public void GetManifestResourceDataAsBase64WithTypeWorks()
         {
-            Assert.AreEqual(Assembly.GetExecutingAssembly().GetManifestResourceDataAsBase64(typeof(AssemblyTests), "Resource1.bin"), "AAECAwQFBgc=", "#1");
-            Assert.AreEqual(Assembly.GetExecutingAssembly().GetManifestResourceDataAsBase64(typeof(AssemblyTests), "Resource2.bin"), "EBESExQV", "#2");
+            Assert.AreEqual("AAECAwQFBgc=", Assembly.GetExecutingAssembly().GetManifestResourceDataAsBase64(typeof(AssemblyTests), "Resource1.bin"), "#1");
+            Assert.AreEqual("EBESExQV", Assembly.GetExecutingAssembly().GetManifestResourceDataAsBase64(typeof(AssemblyTests), "Resource2.bin"), "#2");
             Assert.Null(Assembly.GetExecutingAssembly().GetManifestResourceDataAsBase64(typeof(AssemblyTests), "NonExistent"), "#3");
         }
 
         [Test]
         public void GetManifestResourceDataWithoutTypeWorks()
         {
-            Assert.AreEqual(Assembly.GetExecutingAssembly().GetManifestResourceData("Bridge.ClientTest.Batch1.Reflection.Resource1.bin"), new[] { 0x00, 0x01, 0x02, 0x03, 0x04, 0x05, 0x06, 0x07 }, "#1");
-            Assert.AreEqual(Assembly.GetExecutingAssembly().GetManifestResourceData("Bridge.ClientTest.Batch1.Reflection.Resource2.bin"), new[] { 0x10, 0x11, 0x12, 0x13, 0x14, 0x15 }, "#2");
+            Assert.AreEqual(new[] { 0x00, 0x01, 0x02, 0x03, 0x04, 0x05, 0x06, 0x07 }, Assembly.GetExecutingAssembly().GetManifestResourceData("Bridge.ClientTest.Batch1.Reflection.Resource1.bin"), "#1");
+            Assert.AreEqual(new[] { 0x10, 0x11, 0x12, 0x13, 0x14, 0x15 }, Assembly.GetExecutingAssembly().GetManifestResourceData("Bridge.ClientTest.Batch1.Reflection.Resource2.bin"), "#2");
             Assert.Null(Assembly.GetExecutingAssembly().GetManifestResourceData("NonExistent"), "#3");
         }
 
         [Test]
         public void GetManifestResourceDataWithTypeWorks()
         {
-            Assert.AreEqual(Assembly.GetExecutingAssembly().GetManifestResourceData(typeof(AssemblyTests), "Resource1.bin"), new[] { 0x00, 0x01, 0x02, 0x03, 0x04, 0x05, 0x06, 0x07 }, "#1");
-            Assert.AreEqual(Assembly.GetExecutingAssembly().GetManifestResourceData(typeof(AssemblyTests), "Resource2.bin"), new[] { 0x10, 0x11, 0x12, 0x13, 0x14, 0x15 }, "#2");
+            Assert.AreEqual(new[] { 0x00, 0x01, 0x02, 0x03, 0x04, 0x05, 0x06, 0x07 }, Assembly.GetExecutingAssembly().GetManifestResourceData(typeof(AssemblyTests), "Resource1.bin"), "#1");
+            Assert.AreEqual(new[] { 0x10, 0x11, 0x12, 0x13, 0x14, 0x15 }, Assembly.GetExecutingAssembly().GetManifestResourceData(typeof(AssemblyTests), "Resource2.bin"), "#2");
             Assert.Null(Assembly.GetExecutingAssembly().GetManifestResourceData(typeof(AssemblyTests), "NonExistent"), "#3");
         }
     }
