+++ Tests/Batch4/Batch4.csproj
@@ -54,7 +54,6 @@
     <Compile Include="Exceptions\ContractExceptionTests.cs" />
     <Compile Include="Threading\CancellationTokenTests.cs" />
     <Compile Include="UserDefinedStructTests.cs" />
-    <Compile Include="Reflection\AssemblyTests.cs" />
     <Compile Include="Exceptions\JsErrorExceptionTests.cs" />
     <Compile Include="Reflection\AttributeTests.cs" />
     <Compile Include="SimpleTypes\CharTests.cs" />
